+++ src/Parser/Evaluated/Rule/Natural/KnowledgeCard.php
@@ -29,6 +29,11 @@ class KnowledgeCard implements ParsingRuleInterface
         $data = [
             'title' => function () use ($googleDOM, $node) {
                 $item = $googleDOM->cssQuery('._OKe ._Q1n ._sdf');
+
+                if (!$item->length) {
+                    $item = $googleDOM->cssQuery('.d1rFIf>.kno-ecr-pt>span');
+                }
+
                 return $item->getNodeAt(0)->getNodeValue();
             },
             'shortDescription' => function () use ($googleDOM, $node) {